requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
Pillow>=10.0.0
exifread>=3.0.0
